Kernel/Memory: add VirtualAddress class.
add a VirtualAddress class that replaces raw pointers when dealing with
virtual addresses, this hopefully will help us distinguish between physical
addresses and virtual addresses.

+++ Kernel/Memory/MemoryPage.cpp
@@ -8,7 +8,7 @@ namespace Memory
 
     }
 
-    MemoryPage::MemoryPage(uint64_t base_addr) : MemoryRegion(base_addr, PAGE_SIZE, MULTIBOOT_MEMORY_AVAILABLE)
+    MemoryPage::MemoryPage(uint32_t base_addr) : MemoryRegion(base_addr, PAGE_SIZE, MULTIBOOT_MEMORY_AVAILABLE)
     {
     }
 }
